\name{removeMultipart}
\alias{removeMultipart}
\alias{removeMultipart.MailDocument}
\title{Remove Non-Text Parts From E-Mails}
\description{
  Remove non-text parts from multipart e-mail messages.
}
\usage{
\method{removeMultipart}{MailDocument}(x)
}
\arguments{
  \item{x}{A mail document.}
}
\seealso{
  \code{\link{removeCitation}} to remove e-mail citations, and
  \code{\link{removeSignature}} to remove signature lines from e-mail
  messages.
}
\author{Ingo Feinerer}
